requirements.txt
requests>=2.31.0
pydub>=0.25.1
tenacity>=8.2.3
python-dotenv>=1.0.1
google-api-python-client>=2.121.0
google-auth>=2.28.1
google-auth-oauthlib>=1.2.0
